requirements.txt
Flask==2.3.2
fpdf2==2.7.4